examples/03-integration-examples/aws-documentation-mcp/requirements.txt
# TFrameX and dependencies
-e ../../../

# Additional dependencies for this example
python-dotenv>=1.0.0
asyncio>=3.4.3